+++ app/slackapp/index.js
@@ -1,7 +1,8 @@
 const botkit = require('botkit');
 const { promisify } = require('util');
 const { composeReviewMessage } = require('./review-message');
-const { fillInInteractiveMessage, makeInteractiveMessage } = require('./interactive-message');
+const { fillInInteractiveMessage } = require('./interactive-message');
+const feedSettings = require('./feed-settings');
 
 const setupApp = (slackapp, config, trustpilotApi) => {
   /*
@@ -32,28 +33,13 @@ const setupApp = (slackapp, config, trustpilotApi) => {
   /*
     Internal workings
   */
-  const getTeamFeeds = (team) => team.feeds || [{ channelId: team.incoming_webhook.channel_id, canReply: true }];
-
-  const getChannelFeedSettings = (team, targetChannelId) => {
-    const feeds = getTeamFeeds(team);
-    const channelSettings = feeds.find(({ channelId }) => channelId === targetChannelId);
-    return channelSettings;
-  };
-
-  const getChannelFeedSettingsOrDefault = (team, targetChannelId) => {
-    return getChannelFeedSettings(team, targetChannelId) || { canReply: false };
-  };
-
-  const areSettingsPresentForChannel = (team, targetChannelId) => {
-    return !!getChannelFeedSettings(team, targetChannelId);
-  };
 
   const handleReviewQuery = async (bot, sourceMessage) => {
     let stars = Number(sourceMessage.text.split(' ')[0]);
     stars = isNaN(stars) ? null : stars;
     const team = bot.team_info;
     const businessUnitId = team.businessUnitId;
-    const { canReply } = getChannelFeedSettingsOrDefault(team, sourceMessage.channel);
+    const { canReply } = feedSettings.getChannelFeedSettingsOrDefault(team, sourceMessage.channel);
 
     const lastReview = await trustpilotApi.getLastUnansweredReview({
       stars,
@@ -76,17 +62,12 @@ const setupApp = (slackapp, config, trustpilotApi) => {
 
   const askForReply = (bot, message) => {
     const dialog = bot
-      .createDialog(
-        'Reply to a review',
-        JSON.stringify({
-          dialogType: 'review_reply',
-          originalTs: message.message_ts,
-          reviewId: message.callback_id,
-        }),
-        'Send'
-      )
+      .createDialog('Reply to a review', JSON.stringify({
+        dialogType: 'review_reply',
+        originalTs: message.message_ts,
+        reviewId: message.callback_id,
+      }), 'Send')
       .addTextarea('Your reply', 'reply');
-
     bot.replyWithDialog(message, dialog.asObject(), (err, res) => {
       if (err) {
         console.log(err, res);
@@ -101,170 +82,27 @@ const setupApp = (slackapp, config, trustpilotApi) => {
       channel: message.channel,
       name: 'boom',
     };
-
     try {
       await trustpilotApi.replyToReview({
         reviewId,
         message: message.submission.reply,
       });
-      bot.say(
-        fillInInteractiveMessage({
-          thread_ts: originalTs, // eslint-disable-line camelcase
-          channel: message.channel,
-          attachments: [
-            {
-              author_name: message.raw_message.user.name, // eslint-disable-line camelcase
-              text: message.submission.reply,
-              ts: message.action_ts,
-            },
-          ],
-        })
-      );
+      bot.say(fillInInteractiveMessage({
+        ['thread_ts']: originalTs,
+        channel: message.channel,
+        attachments: [
+          {
+            ['author_name']: message.raw_message.user.name,
+            text: message.submission.reply,
+            ts: message.action_ts,
+          },
+        ],
+      }));
       bot.api.reactions.remove(errorReaction);
     } catch (e) {
       bot.replyPrivateDelayed(message, 'Something went wrong while sending your reply! Please try again shortly.');
       bot.api.reactions.add(errorReaction);
     }
-  };
-
-  const showFeedSettings = (message, bot) => {
-    const team = bot.team_info;
-    const { canReply } = getChannelFeedSettingsOrDefault(team, message.channel);
-    const sourceMessage = {
-      ts: message.message_ts,
-      response_url: message.response_url, // eslint-disable-line camelcase
-      channel: message.channel,
-    };
-    const dialog = bot
-      .createDialog('Review settings', JSON.stringify({ dialogType: 'feed_settings', sourceMessage }), 'Save')
-      .addSelect(
-        'In-channel reply',
-        'replyFeature',
-        canReply ? 'on' : 'off',
-        [
-          { label: 'Allow users to reply to reviews', value: 'on' },
-          {
-            label: 'Do not allow users to reply to reviews', value: 'off',
-          },
-        ]
-      );
-    bot.replyWithDialog(message, dialog.asObject(), (err, res) => {
-      if (err) {
-        console.log(err, res);
-      }
-    });
-  };
-
-  const showFeedSettingsIntroMessage = (message, bot) => {
-    bot.replyInteractiveAsync = bot.replyInteractiveAsync || promisify(bot.replyInteractive);
-    if (areSettingsPresentForChannel(bot.team_info, message.channel)) {
-      return bot.replyInteractiveAsync(
-        message,
-        makeInteractiveMessage({
-          text: 'Manage your review settings for this channel.',
-          actions: [
-            {
-              value: 'delete_feed_settings',
-              text: 'Stop posting reviews',
-              style: 'danger',
-              confirm: {
-                title: 'Are you sure?',
-                text: ' You will no longer see your Trustpilot reviews in this channel.',
-              },
-            },
-            {
-              value: 'open_feed_settings',
-              text: 'Change settings',
-            },
-          ],
-        })
-      );
-    } else {
-      return bot.replyInteractiveAsync(
-        message,
-        makeInteractiveMessage({
-          text:
-            'Get your Trustpilot reviews posted on this channel. ' +
-            'Click the button below to manage your review settings.',
-          actions: [
-            {
-              value: 'open_feed_settings',
-              text: 'Post reviews here',
-              style: 'primary',
-            },
-          ],
-        })
-      );
-    }
-  };
-
-  const upsertFeedSettings = (team, channelId, settings) => {
-    const feeds = team.feeds || [];
-    // Using a Map to upsert the new settings
-    const feedsMap = new Map(feeds.map((f) => [f.channelId, f]));
-    const oldSettings = feedsMap.get(channelId);
-    feedsMap.set(channelId, oldSettings ? { ...oldSettings, ...settings } : settings);
-    team.feeds = [...feedsMap.values()];
-    slackapp.saveTeamAsync = slackapp.saveTeamAsync || promisify(slackapp.saveTeam);
-    return slackapp.saveTeamAsync(team);
-  };
-
-  const handleNewFeedSettings = async (bot, message) => {
-    const {
-      channel: channelId,
-      submission: { replyFeature },
-    } = message;
-    const team = bot.team_info;
-    const businessUnitId = team.businessUnitId;
-    const newSettings = {
-      channelId,
-      businessUnitId,
-      canReply: replyFeature === 'on',
-    };
-    await upsertFeedSettings(team, channelId, newSettings);
-    const privateChannelWarning = channelId.startsWith('G') ? '\nObs! Looks like this is a private channel! ' +
-      'Make sure to */invite @trustpilot* to get reviews posted here!' : '';
-    bot.replyPrivateDelayedAsync = bot.replyPrivateDelayedAsync || promisify(bot.replyPrivateDelayed);
-    if (newSettings.canReply) {
-      await bot.replyPrivateDelayedAsync(
-        message,
-        `All set! Users on this channel can reply to reviews.${privateChannelWarning}`
-      );
-    } else {
-      await bot.replyPrivateDelayedAsync(
-        message,
-        `Settings saved! The reply button is not available to users in this channel.${privateChannelWarning}`
-      );
-    }
-  };
-
-  const deleteFeedSettings = async (message, bot) => {
-    const { channel: channelId } = message;
-    const team = bot.team_info;
-    const feeds = team.feeds || [];
-    // Using a Map to upsert the new settings
-    const feedsMap = new Map(feeds.map((f) => [f.channelId, f]));
-    feedsMap.delete(channelId);
-    team.feeds = [...feedsMap.values()];
-    slackapp.saveTeamAsync = slackapp.saveTeamAsync || promisify(slackapp.saveTeam);
-    await slackapp.saveTeamAsync(team);
-    showFeedSettingsIntroMessage(message, bot);
-  };
-
-  const handleSettingsCommand = async (bot, message) => {
-    bot.api.users.infoAsync = bot.api.users.infoAsync || promisify(bot.api.users.info);
-    const { ok: userOk, user } = await bot.api.users.infoAsync({
-      user: message.user_id,
-    });
-    bot.replyPrivateDelayedAsync = bot.replyPrivateDelayedAsync || promisify(bot.replyPrivateDelayed);
-    if (message.channel_id.startsWith('D')) { // Direct message
-      await bot.replyPrivateDelayedAsync(message, 'Sorry, I can only post your reviews in a proper channel');
-    } else if (userOk && user.is_admin) {
-      await showFeedSettingsIntroMessage(message, bot);
-    } else {
-      await bot.replyPrivateDelayedAsync(message, 'Sorry, only administrators can do that');
-    }
-    return true;
   };
 
   /*
@@ -276,7 +114,7 @@ const setupApp = (slackapp, config, trustpilotApi) => {
     if (/^[1-5] stars?$/i.test(message.text) || /^la(te)?st$/i.test(message.text)) {
       return await handleReviewQuery(bot, message);
     } else if (message.text === 'settings' || message.text === 'feed') {
-      return await handleSettingsCommand(bot, message);
+      return await feedSettings.handleSettingsCommand(bot, message);
     } else {
       return true;
     }
@@ -286,7 +124,7 @@ const setupApp = (slackapp, config, trustpilotApi) => {
     bot.replyAcknowledge();
     const messageAction = message.actions[0].value;
     if (messageAction === 'step_1_write_reply') {
-      const { canReply } = getChannelFeedSettingsOrDefault(bot.team_info, message.channel);
+      const { canReply } = feedSettings.getChannelFeedSettingsOrDefault(bot.team_info, message.channel);
       if (!canReply) {
         bot.replyPublicDelayedAsync = bot.replyPublicDelayedAsync || promisify(bot.replyPublicDelayed);
         await bot.replyPublicDelayedAsync(message, 'Sorry, it’s no longer possible to reply to reviews'
@@ -295,9 +133,9 @@ const setupApp = (slackapp, config, trustpilotApi) => {
         askForReply(bot, message);
       }
     } else if (messageAction === 'open_feed_settings') {
-      showFeedSettings(message, bot);
+      feedSettings.showFeedSettings(message, bot);
     } else if (messageAction === 'delete_feed_settings') {
-      deleteFeedSettings(message, bot);
+      feedSettings.deleteFeedSettings(message, bot, slackapp);
     }
     return true;
   });
@@ -308,9 +146,8 @@ const setupApp = (slackapp, config, trustpilotApi) => {
     if (dialogType === 'review_reply') {
       return handleReply(bot, message);
     } else if (dialogType === 'feed_settings') {
-      await handleNewFeedSettings(bot, message);
-      bot.replyInteractiveAsync = bot.replyInteractiveAsync || promisify(bot.replyInteractive);
-      await showFeedSettingsIntroMessage(sourceMessage, bot);
+      await feedSettings.handleNewFeedSettings(bot, message, slackapp);
+      await feedSettings.showFeedSettingsIntroMessage(sourceMessage, bot);
       return true;
     } else {
       return true;
@@ -327,7 +164,7 @@ const setupApp = (slackapp, config, trustpilotApi) => {
     const bot = slackapp.spawn(team);
     bot.team_info = team; // eslint-disable-line camelcase
     bot.sendAsync = bot.sendAsync || promisify(bot.send);
-    const feeds = getTeamFeeds(team);
+    const feeds = feedSettings.getTeamFeeds(team);
 
     feeds.forEach(async ({ channelId, canReply }) => {
       const message = composeReviewMessage(review, { canReply });
